fixup apigateway and directconnect tests
